// apps/client/src/app/core/swagger/model/systemConfigView.ts
/**
 * NgReports
 * NG Reports API
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface SystemConfigView {
  /**
   * delete reports after
   */
  reportsRetentionInMonths: number;
  /**
   * delete inactive users after
   */
  inactiveUsersRetentionInMonths: number;
  /**
   * enable CORS any
   */
  allowReportsIncomeFromUnknownSources: boolean;
  /**
   * enables AI summary generation
   */
  enableAISummary: boolean;
  /**
   * enabled AI auto summary generation
   */
  enableAIAutoSeverityAssignation: boolean;
  /**
   * enabled AU auto summary generation
   */
  enableAIAutoSummaryGeneration: boolean;
  /**
   * enables including project description to the summary generation context for AI
   */
  summaryGenerationIncludeProjectDescription: boolean;
  /**
   * enables including project environment to the summary generation context for AI
   */
  summaryGenerationIncludeProjectEnvironment: boolean;
  /**
   * enables including report details to the summary generation context for AI
   */
  summaryGenerationIncludeReportDetails: boolean;
  /**
   * enables including report logs to the summary generation context for AI
   */
  summaryGenerationIncludeReportLogs: boolean;
  /**
   * enables including report form data to the summary generation context for AI
   */
  summaryGenerationIncludeReportFormData: boolean;
  /**
   * enables including report environment to the summary generation context for AI
   */
  summaryGenerationIncludeReportEnvironment: boolean;
  /**
   * enables including comments to the summary generation context for AI
   */
  summaryGenerationIncludeComments: boolean;
}